% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-yarra_wq.R
\docType{data}
\name{yarra_wq}
\alias{yarra_wq}
\title{Yarra River Water Quality Dataset}
\format{
A data frame with 876 rows and 16 variables:
\describe{
\item{site}{Monitoring location name}
\item{date}{Date of observation}
\item{ph}{pH level of the water}
\item{temperature}{Water temperature (°C)}
\item{turbidity}{Turbidity (NTU)}
\item{dissolved_oxygen}{Dissolved oxygen concentration (mg/L)}
\item{nitrate}{Nitrate concentration (mg/L)}
\item{...}{Other available environmental indicators}
}
}
\source{
Adapted from your course dataset “yarra_wq_clean.csv”.
}
\usage{
yarra_wq
}
\description{
A cleaned dataset containing water quality measurements from the Yarra River monitoring program.
}
\keyword{datasets}
